feat: refresh file tree after saving files
